+++ src/core/proxy.js
@@ -10,7 +10,7 @@ const requestHandler = async (request, proxy, overrides = {}) => {
         request.continue(); return;
     }
     const cookieHandler = new CookieHandler(request);
-    // Request options for Got accounting for overrides
+    // Request options for GOT accounting for overrides
     const options = {
         cookieJar: await cookieHandler.getCookies(),
         method: overrides.method || request.method(),
@@ -57,42 +57,39 @@ const removeRequestListener = (page, listenerName) => {
     }
 };
 
-// Calls this if request object passed
-const proxyPerRequest = async (request, data) => {
-    let proxy, overrides;
-    // Separate proxy and overrides
-    if (type(data) === "object") {
-        if (Object.keys(data).length !== 0) {
-            proxy = data.proxy;
-            delete data.proxy;
-            overrides = data;
-        }
-    } else {proxy = data}
-    // Skip request if proxy omitted
-    if (proxy) {await requestHandler(request, proxy, overrides)}
-    else {request.continue(overrides)}
-};
+const useProxyPer = {
+    // Call this if request object passed
+    HTTPRequest: async (request, data) => {
+        let proxy, overrides;
+        // Separate proxy and overrides
+        if (type(data) === "object") {
+            if (Object.keys(data).length !== 0) {
+                proxy = data.proxy;
+                delete data.proxy;
+                overrides = data;
+            }
+        } else {proxy = data}
+        // Skip request if proxy omitted
+        if (proxy) {await requestHandler(request, proxy, overrides)}
+        else {request.continue(overrides)}
+    },
 
-// Calls this if page object passed
-const proxyPerPage = async (page, proxy) => {
-    await page.setRequestInterception(true);
-    const listener = "$ppp_request_listener";
-    removeRequestListener(page, listener);
-    const f = {[listener]: async (request) => {
-        await requestHandler(request, proxy);
-    }};
-    if (proxy) {page.on("request", f[listener])}
-    else {await page.setRequestInterception(false)}
-};
+    // Call this if page object passed
+    CDPPage: async (page, proxy) => {
+        await page.setRequestInterception(true);
+        const listener = "$ppp_requestListener";
+        removeRequestListener(page, listener);
+        const f = {[listener]: async (request) => {
+            await requestHandler(request, proxy);
+        }};
+        if (proxy) {page.on("request", f[listener])}
+        else {await page.setRequestInterception(false)}
+    }
+}
 
 // Main function
 const useProxy = async (target, data) => {
-    const targetType = target.constructor.name;
-    if (targetType === "HTTPRequest") {
-        await proxyPerRequest(target, data);
-    } else if (targetType === "Page") {
-        await proxyPerPage(target, data);
-    }
+    useProxyPer[target.constructor.name](target, data);
 };
 
 module.exports = useProxy;