+++ src/index.d.ts
@@ -1,4 +1,4 @@
-export = puppeteer_page_proxy;
+export = useProxy;
 /**
  * **Set a proxy to use in a given page or request.**
  * 
@@ -11,8 +11,8 @@ export = puppeteer_page_proxy;
  * @param page 'Page' or 'Request' object to set a proxy for.
  * @param proxy Proxy to use in the current page. Must begin with a protocol e.g. **http://**, **https://**, **socks://**.
  */
-declare function puppeteer_page_proxy(page: object, proxy: string | object): Promise<any>;
-declare namespace puppeteer_page_proxy {
+declare function useProxy(page: object, proxy: string | object): Promise<any>;
+declare namespace useProxy {
 /**
  * **Request data from a lookupservice.**
  *
@@ -23,9 +23,9 @@ declare namespace puppeteer_page_proxy {
  * console.log(data.ip);
  * ```
  * @param page 'Page' object to execute the request on.
- * @param lookupService External lookup service to request data from. Fetches data from `api64.ipify.org` by default.
+ * @param lookupServiceUrl External lookup service to request data from. Fetches data from `api64.ipify.org` by default.
  * @param isJSON Whether to JSON.parse the received response. Defaults to `true`.
  * @param timeout Time in milliseconds after which the request times out. Defaults to `30000` ms.
  */
-	function lookup(page: object, lookupService?: string, isJSON?: boolean, timeout?: number | string): Promise<any>;
+	function lookup(page: object, lookupServiceUrl?: string, isJSON?: boolean, timeout?: number | string): Promise<any>;
 }